Split out some constants for a few kilobytes more gain
