+++ src/components/sharing/constants.js
@@ -1,0 +1,10 @@
+/**
+ * These are the keys to each permission-lister response object and correspond
+ * to each type of resource that can be shared in Sonora
+ */
+export const TYPE = {
+    DATA: "paths",
+    APPS: "apps",
+    ANALYSES: "analyses",
+    TOOLS: "tools",
+};
